backend/requirements.txt
fastapi==0.135.1
uvicorn==0.34.0
python-multipart==0.0.9
pydantic==2.10.6
pydantic-settings==2.8.1
motor==3.7.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
Pillow==10.4.0
numpy==2.1.3
httpx==0.27.2
google-generativeai
